Create load.py
