+++ ampliwrangler/split.py
@@ -36,7 +36,6 @@ def main(args):
     logger.debug(f'Run ID column name: {args.run_id_column}')
     logger.debug(f'Overwrite existing files: {args.overwrite}')
 
-    set_up_output_directory(args.output_dir, args.overwrite)
     split_manifest_by_run_id(args.input_filepath, args.output_dir, args.run_id_column, args.overwrite)
 
     logger.info(f'{os.path.basename(sys.argv[0])} split: done.')
@@ -104,14 +103,14 @@ def subparse_cli(subparsers, parent_parser: argparse.ArgumentParser = None):
     file_settings = subparser.add_argument_group('Input/output file options')
     run_settings = subparser.add_argument_group('Other params')
 
-    file_settings.add_argument('-i', '--input_filepath', metavar='MANIFEST', required=True,
+    file_settings.add_argument('-i', '--input_filepath', metavar='TSV', required=True,
                                help='The path to the input manifest file. Must match QIIME standards AND have a column '
                                     'with a unique ID for each sequencing run as specified in --run_id_column')
     file_settings.add_argument('-o', '--output_dir', metavar='DIR', required=True,
                                help='The directory where output files (named "manifest_[RUN_ID].tsv") will be saved.')
 
     run_settings.add_argument('-r', '--run_id_column', metavar='STR', required=False, default='run_ID',
-                               help='Name of the column in the input manifest file that contains the unique IDs for '
-                                    'each run. [default: run_ID]')
+                              help='Name of the column in the input manifest file that contains the unique IDs for '
+                                   'each run. [default: run_ID]')
 
     return subparser
